requirements.txt
# =============================================================================
# Health Intelligence App - Unified Requirements
# =============================================================================

# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=1.8.0
python-multipart>=0.0.6

# LangGraph and LangChain (Agent Backend)
langgraph>=0.2.0
langgraph-checkpoint-mongodb>=0.1.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-mongodb>=0.2.0

# MongoDB (Agent Backend)
pymongo>=4.6.0
motor>=3.3.0

# Data Processing (Analytics Backend)
pyarrow>=15.0.0
duckdb>=0.10.0
pandas>=2.2.0
numpy>=1.26.0

# AI/LLM
anthropic>=0.18.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment
python-dotenv>=1.0.0

# HTTP/Async
aiohttp>=3.9.0
httpx>=0.26.0
